# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API backend",
        "description": "Documentação da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empresas": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Lista todas as empresas",
                "operationId": "9239adbd1cfe29bdf0372db2124d0b1f",
                "responses": {
                    "200": {
                        "description": "Lista de Empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empresa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Cria uma nova empresa",
                "operationId": "bea7e1cd70745d9a85563a4cdba64902",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Mostra uma empresa pelo ID",
                "operationId": "7e0d938df2a2305cd42451cc84ec72ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Atualiza uma empresa existente",
                "operationId": "61d6235f805bd62afcf89b24a6bae3aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Deleta uma empresa existente",
                "operationId": "ba97e99dc7744b02f562b861b32e29a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Empresa deletada com sucesso"
                    },
                    "404": {
                        "description": "Empresa não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "title": "Empresa",
                "description": "Detalhes de uma empresa",
                "required": [
                    "id",
                    "nome",
                    "cnpj",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "ID único da empresa",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nome": {
                        "description": "Nome da empresa",
                        "type": "string",
                        "example": "Empresa de Soluções Ltda."
                    },
                    "cnpj": {
                        "description": "CNPJ da empresa",
                        "type": "string",
                        "example": "12.345.678/0001-99"
                    },
                    "perfil": {
                        "description": "Perfil da empresa (opcional)",
                        "type": "string",
                        "example": "Tecnologia e Inovação",
                        "nullable": true
                    },
                    "seguidores": {
                        "description": "Número de seguidores da empresa (opcional)",
                        "type": "integer",
                        "example": 1500,
                        "nullable": true
                    },
                    "email": {
                        "description": "Endereço de e-mail da empresa",
                        "type": "string",
                        "format": "email",
                        "example": "contato@empresasolucao.com"
                    },
                    "telefone": {
                        "description": "Número de telefone da empresa (opcional)",
                        "type": "string",
                        "example": "(11) 98765-4321",
                        "nullable": true
                    },
                    "endereco": {
                        "description": "Endereço completo da empresa (opcional)",
                        "type": "string",
                        "example": "Av. Principal, 1000 - Centro",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp de criação do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Timestamp da última atualização do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "EmpresaCreateRequest": {
                "title": "EmpresaCreateRequest",
                "description": "Dados necessários para criar uma nova empresa",
                "required": [
                    "nome",
                    "cnpj",
                    "email",
                    "senha"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Nova Empresa Ltda."
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "98.765.432/0001-10"
                    },
                    "perfil": {
                        "type": "string",
                        "example": "Consultoria",
                        "nullable": true
                    },
                    "seguidores": {
                        "type": "integer",
                        "example": 50,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "registro@novaempresa.com"
                    },
                    "senha": {
                        "type": "string",
                        "format": "password",
                        "example": "senha_segura123"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "(21) 91234-5678",
                        "nullable": true
                    },
                    "endereco": {
                        "type": "string",
                        "example": "Rua da Inovação, 50 - Bairro Novo",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "EmpresaUpdateRequest": {
                "title": "EmpresaUpdateRequest",
                "description": "Dados opcionais para atualizar uma empresa existente",
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Empresa Atualizada S.A."
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "12.345.678/0001-99"
                    },
                    "perfil": {
                        "type": "string",
                        "example": "Software House",
                        "nullable": true
                    },
                    "seguidores": {
                        "type": "integer",
                        "example": 2000,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "novo.email@empresa.com"
                    },
                    "senha": {
                        "description": "Deixe vazio para não alterar",
                        "type": "string",
                        "format": "password",
                        "example": "nova_senha_456"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "(11) 99887-7665",
                        "nullable": true
                    },
                    "endereco": {
                        "type": "string",
                        "example": "Av. Central, 500 - Distrito Industrial",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": " * @OA\\SecurityScheme(\n * securityScheme=\"sanctum\",\n * type=\"http\",\n * scheme=\"bearer\",\n * bearerFormat=\"Sanctum\"\n * )\n ",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}